api/requirements.txt
uvicorn[standard]
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0

python-dotenv
motor